feat(web): move TechnicalSummary & NewsSentimentPanel to market-intelligence
- Relocated TechnicalSummary and NewsSentimentPanel from bot-analysis to market-intelligence page
- These panels now respond to the asset tab filter (BTC/ETH) and refresh with market data
- Added useMarketSnapshot, useBinanceTicker, useMarketNews, useAgentDecisions hooks to market-intelligence
- Removed unused latestSigmaTechnical computation from bot-analysis
- bot-analysis keeps: CombinedScore → NextDecision → Verdicts → AgentInputSummary
- market-intelligence has: Technical + Sentiment + enriched panels (Fear&Greed, Derivatives, etc.)
