+++ apps/web/src/pages/dashboard/bot-analysis.tsx
@@ -20,14 +20,11 @@ import { useTradingConfigs, useAgentStatus } from '../../hooks/use-trading';
 import { useBinanceTicker } from '../../hooks/use-binance-ticker';
 import { usePlatformMode } from '../../hooks/use-user';
 import {
-  TechnicalSummary,
-  NewsSentimentPanel,
   AgentInputSummary,
   CombinedScoreBanner,
   NextDecisionBanner,
   type NewsAnalysisData,
   type SigmaSentiment,
-  type SigmaTechnical,
 } from '../../components/bot-analysis';
 import { AgentVerdictsBanner } from '../../components/market-intelligence';
 
@@ -154,33 +151,6 @@ export function BotAnalysisPage() {
     return null;
   })();
 
-  // Extract latest SIGMA technical signal from any recent decision
-  const latestSigmaTechnical: SigmaTechnical | null = (() => {
-    for (const d of modeDecisions) {
-      const tech = d.subAgentVerdicts?.find(
-        (v) => v.task === 'technical_signal',
-      );
-      if (tech?.summary) {
-        try {
-          const parsed = JSON.parse(tech.summary);
-          if (parsed.signal) return parsed as SigmaTechnical;
-        } catch {
-          // summary might not be JSON, try extracting signal from text
-          const signalMatch = tech.summary.match(/\b(BUY|SELL|HOLD)\b/i);
-          if (signalMatch) {
-            return {
-              signal: signalMatch[1].toUpperCase(),
-              confidence: 0.5,
-              reasoning: tech.summary,
-              cached: tech.cached,
-            };
-          }
-        }
-      }
-    }
-    return null;
-  })();
-
   const updatedLabel = dataUpdatedAt
     ? t('botAnalysis.updatedAgo', {
         count: Math.round((Date.now() - dataUpdatedAt) / 60_000),
@@ -255,19 +225,6 @@ export function BotAnalysisPage() {
             decisions={modeDecisions}
           />
           <AgentVerdictsBanner />
-          <div className="grid gap-4 lg:grid-cols-2 analysis-section">
-            <TechnicalSummary
-              snapshot={snapshot}
-              livePrice={livePrice}
-              sigmaTechnical={latestSigmaTechnical}
-              enrichedSnapshot={enrichedSnapshot}
-            />
-            <NewsSentimentPanel
-              news={newsItems}
-              sigmaSentiment={latestSigma}
-              enrichedSnapshot={enrichedSnapshot}
-            />
-          </div>
           <AgentInputSummary
             snapshot={snapshot}
             livePrice={livePrice}
